SmartRV/main_service/pyproject.toml
[tool.poetry]
name = "main_service"
version = "0.1.0"
description = ""
authors = ["Dominique Parolin <d.parolin@gmx.net>"]

[tool.poetry.dependencies]
python = ">=3.8, <4"
fastapi = "^0.83.0"
Jinja2 = "^3.0.3"
requests = "^2.27.1"
PyQRCode = "^1.2.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "^0.16.0"
pyserial = "^3.5"
dictdiffer = "^0.9.0"
websockets = "^10.1"
pymodbus = "^2.5.3"
minimalmodbus = "^2.0.1"
setproctitle = "^1.3.1"
pytz = "^2022.7.1"
pypng = "^0.20220715.0"
anyio = "^3.4"

pyyaml = "^6.0.1"
httpx = "^0.24.1"
pytest = "^7.4.1"
pytest-azurepipelines = "^1.0.4"
pytest-asyncio = "^0.23.6"

jsonschema = "^4.19"

[tool.poetry.dev-dependencies]
pyyaml = "^6.0.1"
httpx = "^0.24.1"
pytest = "^7.4.1"
pytest-azurepipelines = "^1.0.4"
jsonschema = "^4.19"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
